tests/wpt/tests/content-security-policy/style-src/style-src-inline-style-with-csstext.html

<!doctype html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="style-src 'self';">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <script>
      var t = async_test("Manipulating cssText should be allowed with 'self'");
      document.addEventListener("securitypolicyviolation", t.unreached_func("Should not trigger a security policy violation"));
    </script>
</head>
<body>
    <div id='log'></div>

    <div id="content">Lorem ipsum</div>

    <script>
      t.step(function() {
        var contentEl = document.getElementById("content");
        contentEl.style.cssText = 'margin-left: 2px;';
        var marginLeftVal = getComputedStyle(contentEl).getPropertyValue('margin-left');
        assert_equals(marginLeftVal, "2px");
        t.done();
      });
    </script>

</body>
</html>
